+++ components/tredly-parse/install.sh
@@ -1,33 +1,33 @@
 #!/usr/local/bin/bash
 
-# Tredly-Host installer
-
 PREFIX="/usr/local"
 MAN=
 BINDIR="${PREFIX}/sbin"
-COMMANDSDIR="${PREFIX}/lib/tredly-host/commands"
-INSTALL=/usr/bin/install
+LIBDIR="${PREFIX}/lib/tredly-parse"
+CONFDIR="${PREFIX}/etc/tredly"
+ACTIONSDIR="${LIBDIR}/actions"
+INSTALL="/usr/bin/install"
 MKDIR="mkdir"
 RM="rm"
 BINMODE="500"
 
-SCRIPTS="tredly-host"
+SCRIPTS="tredly-parse"
 SCRIPTSDIR="${PREFIX}/BINDIR"
 
 function show_help() {
-    echo "Tredly-Host installer"
+    echo "Tredly-Parse installer"
     echo ""
     echo "Usage:"
-    echo "    `basename "$0"` install: install Tredly-Host"
-    echo "    `basename "$0"` uninstall: uninstall Tredly-Host"
-    echo "    `basename "$0"` install clean: remove all previously installed files and install Tredly-Host"
+    echo "    `basename "$0"` install: install Tredly-Parse"
+    echo "    `basename "$0"` uninstall: uninstall Tredly-Parse"
+    echo "    `basename "$0"` install clean: remove all previously installed files and install Tredly-Parse"
 }
 
 # cleans/uninstalls tredly
 function clean() {
     # remove any installed files
     ${RM} -f "${BINDIR}/${SCRIPTS}"
-    ${RM} -f "${COMMANDSDIR}/"*
+    ${RM} -rf "${ACTIONSDIR}/"*
 }
 
 # returns the directory that the files have been downloaded to
@@ -46,33 +46,34 @@ fi
 # where the files are located
 FILESSOURCE=$( get_files_source )
 
-
 # loop over the args, looking for clean first
 for arg in "$@"; do
     if [[ "${arg}" == "clean" ]]; then
-        echo "Cleaning Tredly-Host install"
+        echo "Cleaning Tredly-Parse install"
         clean
     fi
 done
-
 
 # now do it again, but do the install/uninstall
 for arg in "$@"; do
     case "${arg}" in
         install)
-            echo "Installing Tredly-Host..."
-            ${MKDIR} -p "${BINDIR}"
-            ${MKDIR} -p "${COMMANDSDIR}"
-            ${INSTALL} -c -m ${BINMODE} "${FILESSOURCE}/${SCRIPTS}" "${BINDIR}/"
-            ${INSTALL} -c "${FILESSOURCE}/commands/"* "${COMMANDSDIR}"
+            echo "Installing Tredly-Parse..."
 
-            echo "Tredly-Host installed."
+            ${MKDIR} -p "${BINDIR}"
+            ${MKDIR} -p "${LIBDIR}"
+            ${MKDIR} -p "${ACTIONSDIR}"
+            ${INSTALL} -c -m ${BINMODE} "${FILESSOURCE}/${SCRIPTS}" "${BINDIR}/"
+
+            cp -R "${FILESSOURCE}/actions/" "${ACTIONSDIR}"
+
+            echo "Tredly-Parse installed."
             ;;
         uninstall)
-            echo "Uninstalling Tredly-Host..."
+            echo "Uninstalling Tredly-Parse..."
             # run clean to remove the files
             clean
-            echo "Tredly-Host Uninstalled."
+            echo "Tredly-Parse Uninstalled."
             ;;
         clean)
             # do nothing, this is just here to prevent clean being handled as *