+++ components/tredly-validate/actions/validate.py
@@ -17,7 +17,7 @@ from includes.output import *
 
 class ActionValidate:
     def __init__(self, subject, target, identifier, actionArgs):
-        
+
         # check the subject of this action
         if (subject == "container"):
             self.validateContainer(actionArgs['path'])
@@ -29,9 +29,9 @@ class ActionValidate:
     def validateContainer(self, tredlyFilePath):
         # Process the tredlyfile
         builtins.tredlyFile = TredlyFile(tredlyFilePath)
-        
+
         e_header("Validating " + tredlyFilePath)
-        
+
         # validate it
         if (tredlyFile.validate()):
             e_success("Successfully validated " + tredlyFilePath)